move some methods from Logic and Formula to Math
