+++ src/Libraries/CoreNodes/CoreNodes.csproj
@@ -70,9 +70,7 @@
     <Compile Include="Compare.cs" />
     <Compile Include="DateTime.cs" />
     <Compile Include="Files.cs" />
-    <Compile Include="Formula.cs" />
     <Compile Include="ImportExport.cs" />
-    <Compile Include="Logic.cs" />
     <Compile Include="List.cs" />
     <Compile Include="Math.cs" />
     <Compile Include="Object.cs" />
